# Conflicts:
#	Examination/src/components/yuangong.vue
#	Examination/src/router/index.js
#	es-expaper-service/es-expaper-service.iml
#	es-expaper-service/src/main/java/com/trkj/esexpaperservice/dao/Examination.java
#	es-expaper-service/target/classes/com/trkj/esexpaperservice/dao/Examination.class
#	es-user-service/es-user-service.iml

+++ publicservice/src/main/java/com/trkj/user/entity/ExaminationEntity.java
@@ -11,5 +11,5 @@ public class ExaminationEntity implements Serializable {
         private String examname;
         private String examtype;
         private String kssc;
-        private double questionscore;
+        private int questionscore;
 }
